Config: Increase the heartbeat timeout to 15s
Some end points, like the ones under the /feed hierarchy are known to
take longer to return a result, so increase the heartbeat timeout to 15s
to allow them to properly produce a result. To account for the increased
timeout, also raise the memory limit a bit.

Note that this is merely a work-around for the current issue of workers
dying. A proper investigation and optimisations should be done in the
code.

Bug: T189526
Change-Id: Ie621aa3cfad3c04d06e113519f354986129ac321